+++ MekHQ/src/mekhq/campaign/personnel/lifeEvents/FreedomDayAnnouncement.java
@@ -44,7 +44,7 @@ import javax.swing.ImageIcon;
 import megamek.common.annotations.Nullable;
 import megamek.logging.MMLogger;
 import mekhq.campaign.Campaign;
-import mekhq.campaign.CampaignOptions;
+import mekhq.campaign.campaignOptions.CampaignOptions;
 import mekhq.campaign.personnel.Person;
 import mekhq.gui.baseComponents.immersiveDialogs.ImmersiveDialogSimple;
