+++ MekHQ/src/mekhq/gui/InfirmaryTab.java
@@ -57,8 +57,8 @@ import javax.swing.SwingUtilities;
 import megamek.client.ui.util.UIUtil;
 import megamek.common.event.Subscribe;
 import mekhq.MekHQ;
-import mekhq.campaign.CampaignOptions;
 import mekhq.campaign.OptimizeInfirmaryAssignments;
+import mekhq.campaign.campaignOptions.CampaignOptions;
 import mekhq.campaign.event.MedicPoolChangedEvent;
 import mekhq.campaign.event.PersonEvent;
 import mekhq.campaign.event.PersonMedicalAssignmentEvent;
