+++ MekHQ/src/mekhq/gui/dialog/nagDialogs/NagController.java
@@ -39,7 +39,7 @@ import java.util.Collection;
 import java.util.List;
 
 import mekhq.campaign.Campaign;
-import mekhq.campaign.CampaignOptions;
+import mekhq.campaign.campaignOptions.CampaignOptions;
 import mekhq.campaign.finances.Finances;
 import mekhq.campaign.finances.Money;
 import mekhq.campaign.mission.AtBContract;
